Add wall clock

+++ include/day_night.h
@@ -2,7 +2,7 @@
 #define GUARD_DAY_NIGHT_H
 
 
-u8 GetCurrentTimeOfDay(void);
+u16 GetCurrentTimeOfDay(void);
 void LoadClock(void);
 
 #endif // GUARD_DAY_NIGHT_H
